+++ test/opfitd_ms.jl
@@ -4,30 +4,32 @@
 
     @testset "solve_model opfitd (with network inputs): Multi-System Unbalanced case5-case3x2 Without Dist. Generator ACR-ACR" begin
         pm_file = joinpath(dirname(trans_path), "case5_with2loads.m")
-        pmd_file1 = joinpath(dirname(dist_path), "case3_unbalanced_withoutgen.dss")
-        pmd_file2 = joinpath(dirname(dist_path), "case3_unbalanced_withoutgen.dss")
-        # pmd_file1 = joinpath(dirname(dist_path), "case3_unbalanced_withoutgen_mn.dss")
+        # pmd_file1 = joinpath(dirname(dist_path), "case3_unbalanced_withoutgen.dss")
+        # pmd_file2 = joinpath(dirname(dist_path), "case3_unbalanced_withoutgen.dss")
+        pmd_file1 = joinpath(dirname(dist_path), "case3_unbalanced_withoutgen_mn.dss")
         # pmd_file2 = joinpath(dirname(dist_path), "case3_unbalanced_withoutgen_mn.dss")
         # pmd_file2 = joinpath(dirname(dist_path), "case3_balanced.dss")
-        pmd_files = [pmd_file1, pmd_file2]
-        pmitd_file = joinpath(dirname(bound_path), "case5_case3x2.json")
+        pmd_file2 = joinpath(dirname(dist_path), "case3_balanced_mn.dss")
+        pmd_files = [pmd_file1, pmd_file2, pmd_file1]
+        pmitd_file = joinpath(dirname(bound_path), "case5_case3x3.json")
         pmitd_type = NLPowerModelITD{ACRPowerModel, ACRUPowerModel}
 
-        pmitd_data = parse_files(pm_file, pmd_files, pmitd_file, auto_rename=true)
-        # pmitd_data = parse_files(pm_file, pmd_files, pmitd_file; multinetwork=true, auto_rename=true)
+        # pmitd_data = parse_files(pm_file, pmd_files, pmitd_file)
+        # pmitd_data = parse_files(pm_file, pmd_files, pmitd_file; auto_rename=true)
         # pmitd_data = parse_files(pm_file, pmd_files, pmitd_file, multinetwork=true)
+        # pmitd_data = parse_files(pm_file, pmd_files, pmitd_file; multinetwork=true, auto_rename=true)
+
 
         # @info "$(pmitd_data)"
 
-
-        # result = solve_model(pmitd_data, pmitd_type, ipopt, build_opfitd; multinetwork=true)
-        result = solve_model(pmitd_data, pmitd_type, ipopt, build_opfitd)
+        # result = solve_model(pmitd_data, pmitd_type, ipopt, build_opfitd)
+        # result = solve_model(pmitd_data, pmitd_type, ipopt, build_mn_opfitd; multinetwork=true)
 
 
         # result = solve_opfitd(pm_file, pmd_files, pmitd_file, pmitd_type, ipopt; auto_rename=true)
         # result = solve_opfitd(pm_file, pmd_files, pmitd_file, pmitd_type, ipopt)
 
-        # result = solve_mn_opfitd(pm_file, pmd_files, pmitd_file, pmitd_type, ipopt; auto_rename=true)
+        result = solve_mn_opfitd(pm_file, pmd_files, pmitd_file, pmitd_type, ipopt; auto_rename=true)
         # result = solve_mn_opfitd(pm_file, pmd_files, pmitd_file, pmitd_type, ipopt)
 
         @test result["termination_status"] == LOCALLY_SOLVED